UI
Set up the base scenes and UI for those scenes. The main menu will link
to both the instructions and the trial select screen. The trial select
screen will show buttons with the number of trials the user wishes to
play on them. Each cycle is in all by itself 18sec long.  Clicking on a
button from that screen will activate how many cycles the user wishes to
play.